requirement.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2
django-ckeditor==6.1.0
django-countries==7.2.1
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-js-asset==1.2.2
Pillow==8.3.1
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0
